fix compute teacher' services hours
    use native query to call teacher_service_hours
    it returns a double

+++ src/main/java/amu/licence/edt/model/dao/DAOTeacher.java
@@ -8,7 +8,7 @@ import amu.licence.edt.model.beans.Teacher;
 
 public interface DAOTeacher extends DAO<Teacher> {
 
-    public Long computeServiceHours(Teacher t);
+    public Double computeServiceHours(Teacher t);
 
     public List<Teacher> findAvailables(TU tu, Date date, Integer duration);
